requirements.txt
opencv-python
pytesseract
google-genai
pillow
edge-tts
pygame
python-dotenv